requirements.txt
discord.py>=2.3.2
Pillow>=10.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
flask>=2.0.0

